Fix: "header 및 footer 관련 script 분리"
- 스크롤 시 헤더 영역 백그라운드 관련 기능 Header.vue 파일로 이동
- 헤더 백그라운드 관련 css 변경
- 로딩 화면 관련 기능 Footer.vue 파일로 이동
