+++ itty-front/src/main.js
@@ -5,3 +5,5 @@ import router from './router/router.js';
 const app = createApp(App);
 app.use(router);
 app.mount('#app');
+
+"use strict";
